requirements.txt
asyncio
hanlp
pandas
psutil
arrow
httpx
openai
requests
tiktoken
discord.py-self
dotenv